requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.48.0
gradio>=5.50.0
pillow<11.0
numpy>=1.24.0

# Quantization and optimization
bitsandbytes>=0.45.0
accelerate>=0.27.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Evaluation
evaluate>=0.4.1
rouge-score>=0.1.2
bert-score>=0.3.13

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0
requests>=2.31.0